API/requirements.txt
fastapi
uvicorn
cassandra-driver
confluent_kafka